+++ src/reuse/lint.py
@@ -253,6 +253,7 @@ def add_arguments(parser):
 def run(args, out=sys.stdout):
     """List all non-compliant files."""
     project = create_project()
+    project.include_submodules = args.include_submodules
     paths = args.path
     if not paths:
         paths = [project.root]